Grrr. The 'reg' variable in dc_apply_fixup() needs to be a u_int32_t, not a u_int8_t. Pass the conical hat. This should fix certain cardbus 21143 cards that require SROM h0h0magic in order to enable their transceivers.

+++ sys/dev/dc/if_dc.c
@@ -1598,7 +1598,7 @@ static void dc_apply_fixup(sc, media)
 	struct dc_mediainfo	*m;
 	u_int8_t		*p;
 	int			i;
-	u_int8_t		reg;
+	u_int32_t		reg;
 
 	m = sc->dc_mi;
